chore: tests for handleInsufficientBalanceError
